+++ src/app/features/profile/profile.page.ts
@@ -1,4 +1,5 @@
 import { Component } from '@angular/core';
+import { MatDialog } from '@angular/material/dialog';
 import { MatSnackBar } from '@angular/material/snack-bar';
 import { ActivatedRoute, Router } from '@angular/router';
 import { Plugins } from '@capacitor/core';
@@ -6,7 +7,7 @@ import { AlertController } from '@ionic/angular';
 import { TranslocoService } from '@ngneat/transloco';
 import { UntilDestroy, untilDestroyed } from '@ngneat/until-destroy';
 import { defer, forkJoin, iif } from 'rxjs';
-import { catchError, concatMap, concatMapTo } from 'rxjs/operators';
+import { catchError, concatMap, concatMapTo, first } from 'rxjs/operators';
 import { BlockingActionService } from '../../shared/blocking-action/blocking-action.service';
 import { WebCryptoApiSignatureProvider } from '../../shared/collector/signature/web-crypto-api-signature-provider/web-crypto-api-signature-provider.service';
 import { ConfirmAlert } from '../../shared/confirm-alert/confirm-alert.service';
@@ -14,6 +15,7 @@ import { Database } from '../../shared/database/database.service';
 import { DiaBackendAuthService } from '../../shared/dia-backend/auth/dia-backend-auth.service';
 import { DiaBackendWalletService } from '../../shared/dia-backend/wallet/dia-backend-wallet.service';
 import { ErrorService } from '../../shared/error/error.service';
+import { ExportPrivateKeyModalComponent } from '../../shared/export-private-key-modal/export-private-key-modal.component';
 import { MediaStore } from '../../shared/media/media-store/media-store.service';
 import { PreferenceManager } from '../../shared/preference-manager/preference-manager.service';
 
@@ -32,7 +34,8 @@ export class ProfilePage {
   readonly privateKey$ = this.webCryptoApiSignatureProvider.privateKey$;
   readonly phoneVerified$ = this.diaBackendAuthService.phoneVerified$;
   readonly emailVerified$ = this.diaBackendAuthService.emailVerified$;
-  readonly captBalance$ = this.diaBackendWalletService.captBalance$;
+  readonly ethNumBalance$ = this.diaBackendWalletService.ethNumBalance$;
+  readonly bscNumBalance$ = this.diaBackendWalletService.bscNumBalance$;
   readonly networkConnected$ = this.diaBackendWalletService.networkConnected$;
 
   constructor(
@@ -49,7 +52,8 @@ export class ProfilePage {
     private readonly alertController: AlertController,
     private readonly router: Router,
     private readonly route: ActivatedRoute,
-    private readonly diaBackendWalletService: DiaBackendWalletService
+    private readonly diaBackendWalletService: DiaBackendWalletService,
+    private readonly dialog: MatDialog
   ) {}
 
   ionViewWillEnter() {
@@ -112,6 +116,29 @@ export class ProfilePage {
     this.snackBar.open(
       this.translocoService.translate('message.copiedToClipboard')
     );
+  }
+
+  exportPrivateKey() {
+    this.privateKey$
+      .pipe(
+        first(),
+        concatMap(async privateKey => {
+          const result = await this.confirmAlert.present({
+            message: this.translocoService.translate(
+              'message.confirmCopyPrivateKey'
+            ),
+          });
+          if (result)
+            this.dialog.open<ExportPrivateKeyModalComponent>(
+              ExportPrivateKeyModalComponent,
+              {
+                data: { privateKey },
+              }
+            );
+        }),
+        untilDestroyed(this)
+      )
+      .subscribe();
   }
 
   logout() {