+++ src/app/shared/dia-backend/asset/dia-backend-asset-repository.service.ts
@@ -182,7 +182,7 @@ export class DiaBackendAssetRepository {
         }
 
         return this.httpClient.get<PaginatedResponse<DiaBackendAsset>>(
-          `${BASE_URL}/api/v2/assets/`,
+          `${BASE_URL}/api/v3/assets/`,
           { headers, params }
         );
       })
@@ -193,7 +193,7 @@ export class DiaBackendAssetRepository {
     return defer(() => this.authService.getAuthHeaders()).pipe(
       concatMap(headers =>
         this.httpClient.get<DiaBackendAsset>(
-          `${BASE_URL}/api/v2/assets/${id}/`,
+          `${BASE_URL}/api/v3/assets/${id}/`,
           { headers }
         )
       )
@@ -206,7 +206,7 @@ export class DiaBackendAssetRepository {
     return defer(() => this.authService.getAuthHeaders()).pipe(
       concatMap(headers =>
         this.httpClient.post(
-          `${BASE_URL}/api/v2/assets/${id}/download/`,
+          `${BASE_URL}/api/v3/assets/${id}/download/`,
           formData,
           { headers, responseType: 'blob' }
         )
@@ -236,7 +236,7 @@ export class DiaBackendAssetRepository {
     ]).pipe(
       concatMap(([headers, formData]) =>
         this.httpClient.patch<UpdateAssetResponse>(
-          `${BASE_URL}/api/v2/assets/${proof.diaBackendAssetId}/`,
+          `${BASE_URL}/api/v3/assets/${proof.diaBackendAssetId}/`,
           formData,
           { headers }
         )
@@ -247,11 +247,23 @@ export class DiaBackendAssetRepository {
     );
   }
 
+  updateCapture$(id: string, formData: any) {
+    return defer(() => this.authService.getAuthHeaders()).pipe(
+      concatMap(headers =>
+        this.httpClient.patch<UpdateAssetResponse>(
+          `${BASE_URL}/api/v3/assets/${id}/`,
+          formData,
+          { headers }
+        )
+      )
+    );
+  }
+
   removeCaptureById$(id: string) {
     return defer(() => this.authService.getAuthHeaders()).pipe(
       concatMap(headers =>
         this.httpClient.delete<DeleteAssetResponse>(
-          `${BASE_URL}/api/v2/assets/${id}/`,
+          `${BASE_URL}/api/v3/assets/${id}/`,
           { headers }
         )
       ),
@@ -273,7 +285,7 @@ export class DiaBackendAssetRepository {
     return defer(() => this.authService.getAuthHeadersWithApiKey()).pipe(
       concatMap(headers => {
         return this.httpClient.post(
-          `${BASE_URL}/api/v2/assets/${id}/mint/`,
+          `${BASE_URL}/api/v3/assets/${id}/mint/`,
           formData,
           { headers }
         );
@@ -324,6 +336,9 @@ export interface DiaBackendAsset extends Tuple {
   readonly nft_token_uri: string;
   readonly nft_blockchain_name: string;
   readonly nft_contract_address: string;
+  readonly caption: string;
+  readonly post_creation_workflow_id: string;
+  readonly mint_workflow_id: string;
 }
 
 export type AssetDownloadField =